fix(elements): expose TagBullet enum
