+++ src/constants.ts
@@ -4,10 +4,20 @@ export enum Accent {
   TERTIARY = 'TERTIARY'
 }
 
+export enum CoordinatesFormat {
+  DECIMAL_DEGREES = 'DD',
+  DEGREES_MINUTES_DECIMALS = 'DMD',
+  DEGREES_MINUTES_SECONDS = 'DMS'
+}
+
 export enum Size {
   LARGE = 'LARGE',
   NORMAL = 'NORMAL',
   SMALL = 'SMALL'
+}
+
+export enum TagBullet {
+  DISK = 'DISK'
 }
 
 export const WSG84_PROJECTION = 'EPSG:4326'
