--- src/elements/Tag/constants.ts
@@ -1,3 +1,0 @@
-export enum TagBullet {
-  DISK = 'DISK'
-}
