--- src/fields/CoordinatesInput/constants.ts
@@ -1,5 +1,0 @@
-export enum CoordinatesFormat {
-  DECIMAL_DEGREES = 'DD',
-  DEGREES_MINUTES_DECIMALS = 'DMD',
-  DEGREES_MINUTES_SECONDS = 'DMS'
-}
